MirrorDataInProgress moved in DynamoCore project.
