+++ src/Engine/ProtoCore/ProtoCore.csproj
@@ -104,7 +104,6 @@
     <Compile Include="DSASM\HeapObjects.cs" />
     <Compile Include="Lang\InternalAttributes.cs" />
     <Compile Include="Lang\Replication\ArgumentAtLevel.cs" />
-    <Compile Include="Reflection\MirrorDataInProgress.cs" />
     <Compile Include="SyntaxAnalysis\AstVisitor.cs" />
     <Compile Include="AttributeEntry.cs" />
     <Compile Include="BuildStatus.cs" />
